Introduced code that can parse .epub and .mobi files and extract the raw text they contain
I still haven't ensured that management.py works in exactly the way that
I'd like, but I'd like to save that for later. Meanwhile, I'll work on
making sure that all the raw text is ready for tokenization and
embedding. The introduction of this parsing logic for .epub and .mobi
files is one of the initial steps towards that

Note: Loguru had to be downgraded to make this possible.