Decided to implement "partial scraping" by another means. But the solution is incomplete
It works by simply scraping the whole page for text, and then returning
all the text that comes after the last occurence of the provided string.

To ensure that the text we get is as "pure" as possible, I will have to
ensure that text from the latter parts of the text which isn't required
will also be eliminated.