Trying to lay the foundation for the recording of chat history and conversational RAG
One-shot QA has been achieved.
I hope that some of the new code I have written will contribute to the
realisation of conversational RAG.

However, I feel that given what I want to accomplish on the frontend, it
will make sense for me to focus on that side of the application for now.
Hopefully, it will shed some light on exactly how I should code the core
of the conversational system